Move store test util to test-util directory
